+++ game/board.py
@@ -1,4 +1,7 @@
-from game.tiles import Tile
+from game.tiles import *
+from game.player import *
+from game.dictionary import *
+
 TRIPLE_WORD_SCORE = ((0,0), (7, 0), (14,0), (0, 7), (14, 7), (0, 14), (7, 14), (14,14))
 DOUBLE_WORD_SCORE = ((1,1), (2,2), (3,3), (4,4), (1, 13), (2, 12), (3, 11), (4, 10), (13, 1), (12, 2),
                     (11, 3), (10, 4), (13,13), (12, 12), (11,11), (10,10))
@@ -7,75 +10,140 @@ DOUBLE_LETTER_SCORE = ((0, 3), (0,11), (2,6), (2,8), (3,0), (3,7), (3,14), (6,2)
                         (8,2), (8,6), (8,8), (8, 12), (11,0), (11,7), (11,14), (12,6), (12,8), (14, 3), (14, 11))
 class Board:
     def __init__(self):
-        #self.grid = [[Cell(1, '', row=row, col=col) for col in range(15)] for row in range(15)] 
-        self.grid = [
-            [ Cell(1, '') for _ in range(15) ]
-            for _ in range(15)
-        ]
+        self.board = [["   " for i in range(15)] for j in range(15)]
+        self.board[7][7] = " * "
+        self.premium_squares()
+        
 
     def show_board(self):
-        print('\n  |' + ''.join([f' {str(row_index).rjust(2)} ' for row_index in range(15)]))
-        for row_index, row in enumerate(self.grid):
-            print(
-                str(row_index).rjust(2) +
-                '| ' +
-                ' '.join([repr(cell) for cell in row])
-            )
-    def cells_multiplier(self):
+        #board en formato string
+        board_str = "   |  " + "  |  ".join(str(item) for item in range(10)) + "  | " + "  | ".join(str(item) for item in range(10, 15)) + " |"
+        board_str += "\n   _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _\n"
+        board = list(self.board) #copia de board para no modificar la original del init
+        for i in range (len(board)):
+            if i < 10:
+                board[i] = str(i) + "  | " + " | ".join(str(item) for item in board[i]) + " |"
+            if i >= 10:
+                board[i] = str(i) + " | " + " | ".join(str(item) for item in board[i]) + " |"
+        board_str += "\n   |_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _|\n".join(board)
+        board_str += "\n   _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _"
+        return board_str
+
+    def premium_squares(self):
+        #agrega todas las square con multiplicadores al board.
         for coordinate in TRIPLE_WORD_SCORE:
-            self.cell_multiplier(coordinate, "word", 3)
-        for coordinate in DOUBLE_WORD_SCORE:
-            self.cell_multiplier(coordinate, "word", 2)
+            self.board[coordinate[0]][coordinate[1]] = "TWS"
         for coordinate in TRIPLE_LETTER_SCORE:
-            self.cell_multiplier(coordinate, "letter", 3)
+            self.board[coordinate[0]][coordinate[1]] = "TLS"
+        for coordinate in DOUBLE_WORD_SCORE:
+            self.board[coordinate[0]][coordinate[1]] = "DWS"
         for coordinate in DOUBLE_LETTER_SCORE:
-            self.cell_multiplier(coordinate, "letter", 2)
+            self.board[coordinate[0]][coordinate[1]] = "DLS"
 
-    def cell_multiplier(self,coordinate, multiplier_type, multiplier_value):
-        cell = self.grid[coordinate[0]][coordinate[1]]
-        cell.multiplier_type = multiplier_type
-        cell.multiplier = multiplier_value    
-
-    def calculate_word_value(self, word):
-        word_value = 0
-        word_multiplier = 1
-
-        for cell in word:
-            cell_value = cell.calculate_value()
-            word_value += cell_value
-
-            if cell.multiplier_type == 'word':
-                word_multiplier *= cell.multiplier
-                cell.multiplier = 1  # Reinicia el multiplicador de la celda a 1
-        word_value *= word_multiplier
-        return word_value
+        
+    def board_array(self):
+    #Returns del board.
+        return self.board
     
-    def put_words(self, word_list_of_tiles, location, orientation):
-        self.validate_word(word_list_of_tiles, location, orientation)
-        len_word = len(word_list_of_tiles)
-        row, col = location
-        row_increment, col_increment = (0, 1) if orientation == 'H' else (1, 0)
-    # Place the tiles on the board
-        for i in range(len_word):
-            if self.grid[row][col].tile is None: 
-                self.grid[row][col].tile = word_list_of_tiles[i] 
-            row += row_increment
-            col += col_increment
+    def place_word(self, word, location, direction, player):
 
-    def validate_word (self, word, location, orientation):
-        len_word = len(word)
+        # Inicializa la lista de premium_spots
+        premium_spots = []
+        # Verifica la dirección
+        if direction.lower() not in ["right", "down"]:
+            raise ValueError("Dirección no válida. Debe ser 'right' o 'down'.")
 
-        if (orientation == "H" and location[0] + len_word >= 15) or (orientation == "V" and location[1] + len_word >= 15):
-            return False
-        else:
-            return True
+            # Convierte la palabra a mayúsculas
+        word = word.upper()
+
+            # Verifica si la palabra es vacía
+        if not word:
+            raise ValueError("La palabra no puede ser vacía.")
+
+            # Obtiene la longitud de la palabra
+        word_length = len(word)
+
+            # Verifica si la ubicación está dentro de los límites del tablero
+        if not (0 <= location[0] < 15 and 0 <= location[1] < 15):
+            raise ValueError("Ubicación fuera de los límites del tablero.")
+
+            # Comprueba si hay suficiente espacio para colocar la palabra
+        if (direction.lower() == "right" and location[1] + word_length > 15) or (direction.lower() == "down" and location[0] + word_length > 15):
+            raise ValueError("No hay suficiente espacio para colocar la palabra en esa dirección.")
+         
+     # Coloca la palabra en el tablero
+        for i in range(word_length):
+            if direction.lower() == "right":
+                row = location[0]
+                col = location[1] + i
+            elif direction.lower() == "down":
+                row = location[0] + i
+                col = location[1]
+            else:
+                    # Maneja el caso en el que la dirección no es "right" ni "down"
+                raise ValueError("Dirección no válida. Debe ser 'right' o 'down'.")
+
+                
+                # Verifica si la casilla no está vacía
+            if self.board[row][col] != "   ":
+                premium_spots.append((word[i], self.board[row][col]))
+                
+                # Coloca la letra en el tablero
+            self.board[row][col] = " " + word[i] + " "
+
+            # Remueve las fichas del jugador y repone el atril
+        for letter in word:
+            Rack.remove_from_rack(letter)
+
+        Rack.replenish_rack()
+
+
+class Rack:
+    def __init__(self, bag):
+        self.bag = bag
+        self.rack = []
+        self.initialize()
+
+    def initialize(self):
+        #le da sus 7 fichas.
+        for i in range(7):
+            self.add_rack()
+    
+    def add_rack(self):
+        #toma las fichas de la bag y se las da al atril.
+        self.rack.append(self.bag.take(7))
+    
+    def get_rack(self):
+        self.get_letter = Tile.get_letter()
+        #muestra el atril en fomato string.
+        return ", ".join(str(item.get_letter() for item in self.rack)) # join toma una secuencia de elementos y los concatena en una cadena, separándolos con el delimitador especificado (en este caso, la coma seguida de un espacio). convierte cada letra en una cadena utilizando str().
+
+    def get_rack_arr(self):
+       #Devuelve el rack como una matriz de tile
+        return self.rack
+
+    def remove_from_rack(self, tile):
+        #Quita una ficha de la estantería
+        if tile in self.rack:
+           self.rack.remove(tile)
+
+    def get_rack_length(self):
+        #Devuelve el número de fichas que quedan en la estantería.
+        return len(self.rack)
+    
+    def replenish_rack(self):
+        bag = BagTiles()
+        self.get_remaining_tiles = bag.get_remaining_tiles
+        #self.get_remaining_tiles = BagTiles.get_remaining_tiles
+        #Añade fichas a la estantería después de un turno, de forma que la estantería tenga 7 fichas
+        while self.get_rack_length() < 7 and self.get_remaining_tiles() > 0:
+            self.add_rack()
+
 class Cell:
-    def __init__(self, letter=None, multiplier=1, multiplier_type='',multiplier_active=True, tile: Tile = None):
+    def __init__(self,letter = None, multiplier = 1, multiplier_type=None):
         self.multiplier = multiplier
         self.multiplier_type = multiplier_type
         self.letter = letter
-        self.multiplier_active = multiplier_active
-        self.tile = tile
 
     def add_letter(self, letter:Tile):
         self.letter = letter
@@ -84,14 +152,17 @@ class Cell:
         if self.letter is None:
             return 0
         if self.multiplier_type == 'letter':
-            return self.letter.value * self.multiplier
+            value = self.letter.value * self.multiplier
+            self.multiplier_type = None                  
+            return value
         else:
             return self.letter.value
         
-    def __repr__(self):
+    """def __repr__(self):
         if self.letter:
             return repr(self.letter)
         if self.multiplier > 1:
             return f'{"W" if self.multiplier_type == "word" else "L"}x{self.multiplier}'
         else:
             return '   '
+"""