+++ game/main.py
@@ -1,8 +1,0 @@
-from game.scrabbleCli import ScrabbleCli
-
-def main():
-    client = ScrabbleCli()
-    client.client()
-
-if __name__ == '__main__':
-    main()